move server controller methods into separate files
