+++ packages/commonwealth/server/routes/threads/update_thread_handler.ts
@@ -24,7 +24,7 @@ export const updateThreadHandler = async (
 ) => {
   const { user, address, chain } = req;
   const { id } = req.params;
-  const { body, title, kind, stage, url } = req.body;
+  const { body, title, stage, url } = req.body;
 
   const threadId = parseInt(id, 10) || 0;
   if (!threadId) {
@@ -39,17 +39,16 @@ export const updateThreadHandler = async (
   }
 
   const [updatedThread, notificationOptions] =
-    await controllers.threads.updateThread(
+    await controllers.threads.updateThread({
       user,
       address,
       chain,
       threadId,
       title,
       body,
-      kind,
       stage,
-      url
-    );
+      url,
+    });
 
   for (const n of notificationOptions) {
     controllers.notifications.emit(n).catch(console.error);
